+++ src/main/java/com/zepben/ewb/datastores/energy/EwbEnergyProfileStoreReindexer.java
@@ -13,7 +13,7 @@ import com.zepben.blobstore.BlobReader;
 import com.zepben.blobstore.BlobStoreException;
 import com.zepben.blobstore.itemwrappers.ByDateBlobReaderProvider;
 import com.zepben.energy.datastore.blobstore.indexing.BlobDateRangeIndex;
-import com.zepben.ewb.filepaths.EwbDataFilePaths;
+import com.zepben.evolve.database.paths.EwbDataFilePaths;
 import kotlin.Unit;
 
 import java.io.IOException;