+++ src/test/java/com/zepben/ewb/datastores/energy/SqliteEwbEnergyProfileStoreTest.java
@@ -15,7 +15,7 @@ import com.zepben.energy.model.Channel;
 import com.zepben.energy.model.EnergyProfile;
 import com.zepben.energy.model.IdDateRange;
 import com.zepben.energy.model.Readings;
-import com.zepben.ewb.filepaths.EwbDataFilePaths;
+import com.zepben.evolve.database.paths.EwbDataFilePaths;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;